Added interpreter func, moved parser
